feat(swarms): product-scoped client management + journey chips (PR 3)
Swarms can now create and configure its OWN clients without minting a
Chatbox/publish surface, reusing the shared host editor and a product-
neutral server-group picker.

- SwarmsTab: add a "Journeys | Clients" sub-nav. Clients mounts the new
  SwarmHostsPanel; Journeys keeps the persona→journey view.
- SwarmHostsPanel (new): lists journeys-owned + untagged clients (hides
  chatbox/suite-owned); per row model · N servers · computer chips; Edit
  deep-links into the ONE shared host editor (Connect); "Apply server group"
  snapshot-copies a group's servers INTO the host config via hosts:updateHost
  (full replace, optional cleared — never chatboxes:setChatboxServers); New
  client creates a standalone host (owner:'journeys'). Delete surfaces the
  backend's active-journey rejection.
- Role gating (TL must-fix): New/Edit/Apply/Delete are admin-only; threaded
  via canManageHosts(viewerRole) from SwarmsRoute. Members still view + run.
- Server-group picker extracted to product-neutral
  components/hosts/ServerGroupPicker (moved from evals/server-attachment-picker,
  which is now a re-export shim so eval/chat-v2 call sites are unchanged).
- CreateHostDialog: optional owner?: 'journeys' prop (every existing mount
  omits it → byte-identical).
- Journey host-picker chips now show model · servers · computer.
- Tests: panel filtering, chip rendering, admin/member role gating, create
  passes owner:'journeys', apply issues updateHost with servers replaced.

Depends on PR 1 (backend ownerScope) deployed + PR 2 (useClients types).
